crm_agent/requirements.txt
django==5.1.2
django-ninja==1.1.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
email-validator==2.3.0
pandas==2.2.2
openpyxl==3.1.5
pypdf==4.3.1
pillow==10.4.0
pdf2image==1.17.0
pytesseract==0.3.10
chromadb==0.5.5
sentence-transformers==3.0.1
tiktoken==0.8.0